EA3131: Rename files for better conformance to naming conventions
