+++ nuttx/configs/ea3131/src/lpc31_spi.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/ea3131/src/up_spi.c
- * arch/arm/src/board/up_spi.c
+ * configs/ea3131/src/lpc31_spi.c
  *
  *   Copyright (C) 2009-2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -139,4 +138,3 @@ uint8_t lpc31_spistatus(FAR struct spi_dev_s *dev, enum spi_dev_e devid)
 
 #endif /* 0 */
 #endif /* CONFIG_LPC31_SPI  */
-
